Issue #62: extract shared hook install/uninstall utilities
Move getGitBash, toBashPath, installHooks, and uninstallHooks from
projects.ts into src/server/utils/hook-installer.ts. Update system.ts
factory-reset to use the shared uninstallHooks instead of its ~15-line
inline duplicate.
