+++ src/server/routes/projects.ts
@@ -6,7 +6,6 @@
 // =============================================================================
 
 import type {
-  FastifyBaseLogger,
   FastifyInstance,
   FastifyPluginCallback,
   FastifyRequest,
@@ -20,6 +19,7 @@ import { getTeamManager } from '../services/team-manager.js';
 import { getIssueFetcher } from '../services/issue-fetcher.js';
 import { getCleanupPreview, executeCleanup } from '../services/cleanup.js';
 import { sseBroker } from '../services/sse-broker.js';
+import { installHooks, uninstallHooks } from '../utils/hook-installer.js';
 import config from '../config.js';
 import type { ProjectStatus, InstallStatus, InstallFileStatus, CleanupPreview, CleanupResult } from '../../shared/types.js';
 
@@ -65,37 +65,6 @@ function normalizePath(p: string): string {
 }
 
 /**
- * Find Git Bash executable on Windows.
- * Avoids WSL's bash which can't handle Windows paths.
- */
-let _gitBash: string | null = null;
-function getGitBash(): string {
-  if (_gitBash) return _gitBash;
-  try {
-    // git --exec-path returns e.g. "C:/Git/scm/libexec/git-core"
-    // Git bash is at {git_root}/usr/bin/bash.exe
-    const execPath = execSync('git --exec-path', { encoding: 'utf-8', stdio: 'pipe' }).trim();
-    const gitRoot = path.resolve(execPath, '..', '..');
-    const bashPath = path.join(gitRoot, 'usr', 'bin', 'bash.exe');
-    if (fs.existsSync(bashPath)) {
-      _gitBash = bashPath;
-      return _gitBash;
-    }
-  } catch {
-    // fallback
-  }
-  _gitBash = 'bash'; // hope for the best
-  return _gitBash;
-}
-
-/**
- * Convert a Windows path to a bash-safe format with forward slashes.
- */
-function toBashPath(p: string): string {
-  return p.replace(/\\/g, '/');
-}
-
-/**
  * Check if a directory is a git repository.
  */
 function isGitRepo(dirPath: string): boolean {
@@ -124,63 +93,6 @@ function detectGithubRepo(dirPath: string): string | null {
     return result || null;
   } catch {
     return null;
-  }
-}
-
-/**
- * Install Fleet Commander hooks into a project repo.
- * Returns { ok, stdout, stderr } so callers can log / surface errors.
- */
-function installHooks(repoPath: string, logger: FastifyBaseLogger): { ok: boolean; stdout: string; stderr: string } {
-  const fail = (msg: string) => ({ ok: false, stdout: '', stderr: msg });
-
-  const scriptPath = path.join(config.fleetCommanderRoot, 'scripts', 'install.sh');
-  if (!fs.existsSync(scriptPath)) {
-    return fail(`install.sh not found at ${scriptPath}`);
-  }
-
-  // On Windows, convert paths to MSYS2 format: C:/foo → /c/foo
-  const cmd = process.platform === 'win32'
-    ? `"${getGitBash()}" "${toBashPath(scriptPath)}" "${toBashPath(repoPath)}"`
-    : `"${scriptPath}" "${repoPath}"`;
-
-  try {
-    const stdout = execSync(cmd, { encoding: 'utf-8', stdio: 'pipe', timeout: 30000 });
-    return { ok: true, stdout, stderr: '' };
-  } catch (err: unknown) {
-    const e = err as { stdout?: string; stderr?: string; message?: string; status?: number };
-    const stdout = e.stdout ?? '';
-    const stderr = e.stderr ?? '';
-    logger.error(
-      `[installHooks] Failed for ${repoPath} (exit ${e.status ?? '?'}):\n` +
-      `  cmd: ${cmd}\n` +
-      `  stderr: ${stderr.trim()}\n` +
-      `  stdout: ${stdout.trim()}`,
-    );
-    return { ok: false, stdout, stderr: stderr || e.message || 'unknown error' };
-  }
-}
-
-/**
- * Uninstall Fleet Commander hooks from a project repo.
- */
-function uninstallHooks(repoPath: string, logger: FastifyBaseLogger): void {
-  try {
-    const scriptPath = path.join(config.fleetCommanderRoot, 'scripts', 'uninstall.sh');
-    if (!fs.existsSync(scriptPath)) {
-      return;
-    }
-
-    const cmd = process.platform === 'win32'
-      ? `"${getGitBash()}" "${toBashPath(scriptPath)}" "${toBashPath(repoPath)}"`
-      : `"${scriptPath}" "${repoPath}"`;
-
-    execSync(cmd, { encoding: 'utf-8', stdio: 'pipe', timeout: 30000 });
-  } catch (err) {
-    // Non-fatal — log but don't block project deletion
-    logger.error(
-      `[uninstallHooks] Failed to uninstall from ${repoPath}: ${err instanceof Error ? err.message : String(err)}`,
-    );
   }
 }
